requirements.txt
Flask
reportlab
Pillow
django
djangorestframework
django-cors-headers
